+++ pandapower/results_branch.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 
 from pandapower.auxiliary import _sum_by_group
-from pandapower.idx_brch import F_BUS, T_BUS, PF, QF, PT, QT
+from pandapower.idx_brch import F_BUS, T_BUS, PF, QF, PT, QT, BR_R
 from pandapower.idx_bus import BASE_KV
 
 
@@ -82,6 +82,15 @@ def _get_line_results(net, ppc, i_ft):
     res_line_df["i_to_ka"].values[:] = i_to_ka
     res_line_df["i_ka"].values[:] = i_ka
     res_line_df["loading_percent"].values[:] = i_ka / i_max * 100
+
+    # if consider_line_temperature, add resulting r_ohm_per_km to net.res_line
+    if net["_options"]["consider_line_temperature"]:
+        from_bus = ppc['branch'][f:t, F_BUS].real.astype(np.int64)
+        base_kv = ppc["bus"][from_bus, BASE_KV]
+        baseR = np.square(base_kv) / net.sn_kva * 1e3
+        length_km = line_df.length_km.values
+        parallel = line_df.parallel.values
+        res_line_df["r_ohm_per_km"] = ppc["branch"][f:t, BR_R].real / length_km * baseR * parallel
 
 
 def _get_trafo_results(net, ppc, s_ft, i_ft):