+++ pandapower/run.py
@@ -36,10 +36,11 @@ def set_user_pf_options(net, overwrite=False, **kwargs):
     :return: None
     """
     standard_parameters = ['calculate_voltage_angles', 'trafo_model', 'check_connectivity', 'mode',
-                           'copy_constraints_to_ppc', 'r_switch', 'init', 'enforce_q_lims',
-                           'recycle', 'voltage_depend_loads', 'delta', 'tolerance_kva',
-                           'trafo_loading', 'numba', 'ac', 'algorithm', 'max_iteration',
-                           'trafo3w_losses', 'init_vm_pu', 'init_va_degree']
+                           'copy_constraints_to_ppc', 'r_switch', 'enforce_q_lims', 'recycle',
+                           'voltage_depend_loads', 'consider_line_temperature', 'delta',
+                           'trafo3w_losses', 'init_vm_pu', 'init_va_degree',  'init_results',
+                           'tolerance_kva', 'trafo_loading', 'numba', 'ac', 'algorithm',
+                           'max_iteration', 'v_debug']
 
     if overwrite or 'user_pf_options' not in net.keys():
         net['user_pf_options'] = dict()
@@ -82,7 +83,7 @@ def _passed_runpp_parameters(local_parameters):
 def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto",
           max_iteration="auto", tolerance_kva=1e-5, trafo_model="t",
           trafo_loading="current", enforce_q_lims=False, check_connectivity=True,
-          voltage_depend_loads=True, **kwargs):
+          voltage_depend_loads=True, consider_line_temperature=False, **kwargs):
     """
     Runs a power flow
 
@@ -166,6 +167,11 @@ def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto",
             If check finds unsupplied buses, they are set out of service in the ppc
 
         **voltage_depend_loads** (bool, True) - consideration of voltage-dependent loads. If False, net.load.const_z_percent and net.load.const_i_percent are not considered, i.e. net.load.p_kw and net.load.q_kvar are considered as constant-power loads.
+
+        **consider_line_temperature** (bool, False) - adjustment of line impedance based on provided
+            line temperature. If True, net.line must contain a column "temperature_degree_celsius".
+            The temperature dependency coefficient alpha must be provided in the net.line.alpha
+            column, otherwise the default value of 0.004 is used
 
 
         **KWARGS:
@@ -283,7 +289,8 @@ def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto",
                      mode=mode, copy_constraints_to_ppc=copy_constraints_to_ppc,
                      r_switch=r_switch, init_vm_pu=init_vm_pu, init_va_degree=init_va_degree,
                      enforce_q_lims=enforce_q_lims, recycle=recycle,
-                     voltage_depend_loads=voltage_depend_loads, delta=delta_q,
+                     voltage_depend_loads=voltage_depend_loads,
+                     consider_line_temperature=consider_line_temperature, delta=delta_q,
                      trafo3w_losses=trafo3w_losses)
     _add_pf_options(net, tolerance_kva=tolerance_kva, trafo_loading=trafo_loading,
                     numba=numba, ac=ac, algorithm=algorithm, max_iteration=max_iteration,
@@ -364,7 +371,7 @@ def rundcpp(net, trafo_model="t", trafo_loading="current", recycle=None, check_c
 
 def runopp(net, verbose=False, calculate_voltage_angles=False, check_connectivity=False,
            suppress_warnings=True, r_switch=0.0, delta=1e-10, init="flat", numba=True,
-           trafo3w_losses="hv", **kwargs):
+           trafo3w_losses="hv", consider_line_temperature=False, **kwargs):
     """
     Runs the  pandapower Optimal Power Flow.
     Flexibilities, constraints and cost parameters are defined in the pandapower element tables.
@@ -415,6 +422,11 @@ def runopp(net, verbose=False, calculate_voltage_angles=False, check_connectivit
             "pf": a power flow is executed prior to the opf and the pf solution is the starting vector. This may improve
             convergence, but takes a longer runtime (which are probably neglectible for opf calculations)
 
+        **consider_line_temperature** (bool, False) - adjustment of line impedance based on provided
+            line temperature. If True, net.line must contain a column "temperature_degree_celsius".
+            The temperature dependency coefficient alpha must be provided in the net.line.alpha
+            column, otherwise the default value of 0.004 is used
+
          **kwargs** - Pypower / Matpower keyword arguments: - OPF_VIOLATION (5e-6) constraint violation tolerance
                                                             - PDIPM_COSTTOL (1e-6) optimality tolerance
                                                             - PDIPM_GRADTOL (1e-6) gradient tolerance
@@ -441,7 +453,9 @@ def runopp(net, verbose=False, calculate_voltage_angles=False, check_connectivit
                      mode=mode, copy_constraints_to_ppc=copy_constraints_to_ppc,
                      r_switch=r_switch, init_vm_pu=init, init_va_degree=init,
                      enforce_q_lims=enforce_q_lims, recycle=recycle,
-                     voltage_depend_loads=False, delta=delta, trafo3w_losses=trafo3w_losses)
+                     voltage_depend_loads=False,
+                     consider_line_temperature=consider_line_temperature, delta=delta,
+                     trafo3w_losses=trafo3w_losses)
     _add_opf_options(net, trafo_loading=trafo_loading, ac=ac, init=init, numba=numba)
     _check_bus_index_and_print_warning_if_high(net)
     _check_gen_index_and_print_warning_if_high(net)